# Server/src/main/java/fr/makizart/common/api/api.yaml
openapi: 3.0.0
info:
  title: Admin API
  version: 1.0.0
  description: API for managing projects and user authentication

paths:
  /admin/projects/create/project/:
    post:
      summary: Create a new project
      description: Create a new project within the admin account.
      parameters:
        - in: query
          name: name
          required: true
          schema:
            type: string
          description: The name of the new project.
      responses:
        200:
          description: Project successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/created_id'
      security:
        - SpringOAuth2: [admin]

  /admin/projects/{project_id}/create/resource/:
    post:
      summary: Create a resource in a project
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: integer
          description: The ID of the project where the new resource will be created
        - in: query
          name: resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/uploadResource"

      responses:
        200:
          description: resource created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/created_id'
        404:
          description: Id not found
        409:
          description: If resource already exist
        5XX:
          description: Unexpected error
      security:
        - SpringOAuth2: [admin]

  /admin/projects/resources/{resource_id}/image/:
    put:
      summary: Upload a image on a resource
      description: Upload an image on a resource, override if already exist.
      parameters:
        - in: path
          name: resource_id
          required: true
          schema:
            type: integer
          description: The ID of the project where to upload the image.
        - in: query
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/created_id'
          description: Image uploaded successfully.
        400:
          description: Invalid name or unsupported media
        404:
          description: Resource not found. The specified resource ID  does not exist.
        5XX:
          description: Unexpected error
      security:
        - SpringOAuth2: [admin]

  /admin/projects/resources/{resource_id}/trackedImage/:
    put:
      summary: Upload a tracked image on a resource
      description: Upload an tracked image on a resource, override if already exist.
      parameters:
        - in: path
          name: resource_id
          required: true
          schema:
            type: integer
          description: The ID of the resource where to upload the tracked image.
        - in: query
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              format: binary
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/created_id'
          description: Tracked image uploaded successfully.
        400:
          description: Invalid name or unsupported media
        404:
          description: Resource not found. The specified resource ID  does not exist.
        5XX:
          description: Unexpected error
      security:
        - SpringOAuth2: [admin]

  /admin/projects/resources/{resource_id}/markers/:
    put:
      summary: Upload markers on a resource
      description: Upload markers on a resource, override if already exist.
      parameters:
        - in: path
          name: resource_id
          required: true
          schema:
            type: integer
          description: The ID of the resource where to upload the markers.
        - in: query
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          marker1/iset:
            schema:
              type: string
              format: binary
          marker2/fset:
            schema:
              type: string
              format: binary
          marker3/fset3:
            schema:
              type: string
              format: binary
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/created_id'
          description: Markers uploaded successfully.
        400:
          description: Invalid name or unsupported media
        404:
          description: Resource not found. The specified resource ID  does not exist.
        5XX:
          description: Unexpected error
      security:
        - SpringOAuth2: [admin]

  /admin/projects/resources/{resource_id}/sound/:
    put:
      summary: Upload sound on a resource
      description: Upload sound on a resource, override if already exist.
      parameters:
        - in: path
          name: resource_id
          required: true
          schema:
            type: string
          description: The ID of the resource where to upload the sound.
        - in: query
          name: name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          audio/mp3:
            schema:
              type: string
              format: binary

      responses:
        200:
          description: Tracked image uploaded successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/created_id'
        400:
          description: Invalid name or unsupported media
        404:
          description: Resource not found. The specified resource ID  does not exist.
        5XX:
          description: Unexpected error
      security:
        - SpringOAuth2: [ admin ]

  /admin/projects/resources/{resource_id}/video/:
    put:
      summary: Upload video on a resource
      description: Upload video on a resource
      parameters:
        - in: path
          name: resource_id
          required: true
          schema:
            type: string
          description: The ID of the resource where to upload the video.
        - in: query
          name: name
          required: true
          schema:
            type: string
        - in: query
          name: url
          required: true
          schema:
            type: string
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/created_id'
          description: Tracked image uploaded successfully.
        400:
          description: Invalid name or unsupported media
        404:
          description: Resource not found. The specified resource ID  does not exist.
        5XX:
          description: Unexpected error
      security:
        - SpringOAuth2: [ admin ]
        
  /public/projects/{project_id}/rename:
    put:
      summary: Rename a project
      description: Rename a project by specifying the new project id.
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
          description: The ID of the project.
        - in: query
          name: new_name
          required: true
          schema:
            type: string
          description: The new name of the project.
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/created_id'
          description: Rename successful.
        404:
          description: Unauthorized or not found.
        5XX:
          description: Unexpected error
      security:
        - SpringOAuth2: [ admin ]
        
        
  /public/projects/resources/{resource_id}/rename:
    put:
      summary: Rename a resource
      description: Rename a resource by specifying the new resource id.
      parameters:
        - in: path
          name: resource_id
          required: true
          schema:
            type: string
          description: The ID of the resource.
        - in: query
          name: new_name
          required: true
          schema:
            type: string
          description: The new name of the resource.
      responses:
        200:
          description: Rename successful.
        404:
          description: Unauthorized or not found.
        5XX:
          description: Unexpected error
      security:
        - SpringOAuth2: [ admin ]

  /admin/projects/resources/medias/{media_id}/rename:
    put:
      summary: Rename media
      description: Rename a media by specifying the media ID.
      parameters:
        - in: path
          name: media_id
          required: true
          schema:
            type: string
          description: The ID of the media.
      responses:
        200:
          description: Rename successfully.
        400:
          description: Project not found or invalid permission.
        5XX:
          description: Unexpected error
      security:
        - SpringOAuth2: [ admin ]
        
  /public/projects/{project_id}/delete:
    delete:
      summary: Delete a project
      description: Delete a project by specifying the project id.
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
          description: The ID of the project.
      responses:
        200:
          description: Rename successful.
        404:
          description: Unauthorized or not found.
        5XX:
          description: Unexpected error
      security:
        - SpringOAuth2: [ admin ]
        
  /public/projects/resources/{resource_id}/delete:
    delete:
      summary: Delete a resource
      description: Delete a resource by specifying the resource id.
      parameters:
        - in: path
          name: resource_id
          required: true
          schema:
            type: string
          description: The ID of the resource.
      responses:
        200:
          description: Rename successful.
        404:
          description: Unauthorized or not found.
        5XX:
          description: Unexpected error.
      security:
        - SpringOAuth2: [ admin ]

  /public/projects/:
    get:
      parameters:
        - in: query
          name: page
          required: true
          schema:
            type: integer
        - in: query
          name: size
          required: true
          schema:
            type: integer
      summary: Get a list of projects
      description: Returns a list of public projects.
      responses:
        400:
          description: Invalid parameter,
        200:
          description: List of projects as string
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        name:
                          type: string
                        creationDate:
                          type: string
  
  /public/projects/{project_id}/resources/:
    get:
      summary: Get a list of resources in project
      description: Returns a list of resource in a projects specified by ID.
      parameters:
        - in: path
          name: project_id
          required: true
          schema:
            type: string
      responses:
          200:
            description: All resources in a project
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    resources:
                      type: array
                      items:
                        type: object
                        properties:
                          id:
                            type: integer
                          name:
                            type: string
                          creationDate:
                            type: string
                          thumbnailId:
                            description: "id of thumbnail image"
                            type: string
                          markersId:
                            description: "id of markers"
                            type: string
                          imageAssetId:
                            description: "id of image asset"
                            type: string
                          videoAssetId:
                            description: "id of video asset"
                            type: string
                          soundAssetId:
                            description: "id of soundAssetId"
                            type: string
          404:
            description: Project not found
            
  /public/projects/resources/{resource_id}/:
    get:
      summary: Get a resource
      parameters:
        - name: resource_id
          in: path
          description: ID of the resource to be accessed
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Return all project element list
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  creationDate:
                    type: string
                  thumbnail_id:
                    type: integer
                  marker1_id:
                    type: integer
                  marker2_id:
                    type: integer
                  marker3_id:
                    type: integer
                  image_id:
                    type: integer
                  sound_id:
                    type: integer
                  video_id:
                    type: integer
        404:
          description: resource not found
        5XX:
          description: Bad request



  /public/storage/:
    get:
      summary: Get storage information
      description: Get the free space and max space on the disk (in Mb)
      responses:
        200:
          description: Storage information as string
          content:
            application/json:
              schema:
                type: object
                properties:
                  used:
                    type: integer
                  total:
                    type: integer

components:
  schemas:
    uploadResource:
      type: object
      properties:
        name:
          type: string
        marker1:
          type: string
          description: Base64 pathToMarker1
        marker2:
          type: string
          description: Base64 pathToMarker2
        marker3:
          type: string
          description: Base64 pathToMarker3
        thumbnail:
          type: string
          description: Base64 thumbnail
        imageAsset:
          type: string
          description: Base64 imageAsset
        videoAsset:
          type: string
          description: video url
        soundAsset:
          type: string
          description: Base64 soundAsset
    resourceAsJSON:
      type: object
      properties:
        name:
          type: string
        pathToMarker1Id:
          type: string
          description: ID of pathToMarker1
        pathToMarker2Id:
          type: string
          description: ID of pathToMarker2
        pathToMarker3Id:
          type: string
          description: ID of pathToMarker3
        thumbnailId:
          type: string
          description: ID of thumbnail
        imageAssetId:
          type: string
          description: ID of imageAsset
        videoAssetId:
          type: string
          description: ID of videoAsset
        soundAssetId:
          type: string
          description: ID of soundAsset
    created_id:
      type: object
      properties:
        application/json:
            type: object
            properties:
              id:
                type: integer
                description: id of the created element

  securitySchemes:
    SpringOAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: "http://localhost:9000/oauth2/authorize"
          tokenUrl: "http://localhost:9000/oauth2/token"
          scopes:
            admin: Grants access to admin read/write operations